Some refactoring and added validation to buffer and texture copies
Former-commit-id: 2b1415a102ece81e3e82dddb7d4104ac03edfdfa [formerly 396d1ef60edc798311272e3af2f8d3fcd64a68ab]
Former-commit-id: f8107f0e677481f7efcea1cdf1ff53b332cb8c2a

+++ fancy_core/VkPrerequisites.h
@@ -30,7 +30,7 @@ namespace Fancy
     if (kVkImplementationDebugLevel > VK_IMPLEMENTATION_DEBUG_LEVEL_NONE)
       Log("Missing Vulkan Implementation: %s", aFunction);
 
-    if (kVkImplementationDebugLevel == VK_IMPLEMENTATION_DEBUG_LEVEL_ASSERT)
+    if ((uint) kVkImplementationDebugLevel == (uint) VK_IMPLEMENTATION_DEBUG_LEVEL_ASSERT)
       throw;
   }
